requirements.txt
pytorch-lightning==1.1.7
hydra-core==1.0.6
torchtext